feat(login): moved the login to the auth module
